+++ src/renderer/greyscaleLUT.glsl
@@ -1,4 +1,4 @@
-precision mediump float;
+precision highp float;
 uniform bool u_invert;
 uniform sampler2D u_texture;
 uniform vec2 u_resolution;
@@ -16,13 +16,13 @@ void main() {
   uv.y = 1.0 - uv.y;
 
   float grey = greyscale(texture2D(u_texture, uv));
-  
+
   float lutPos = (max(u_firstInputValue, grey) - u_firstInputValue);
   grey = greyscale(texture2D(u_lutTexture, vec2(lutPos / u_lutWidth, 0.5))) / u_maxValue;
-  
+
   if (u_invert) {
     grey = 1.0 - grey;
   }
-  
+
   gl_FragColor = vec4(grey, grey, grey, 1);
 }
